pyproject.toml
[project]
name = "pro-plan-api"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.2",
    "psycopg[binary]>=3.3.3",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "sqlalchemy-utils>=0.42.1",
    "ty>=0.0.24",
]

[tool.pytest.ini_options]
anyio_mode = "auto"

[tool.ruff.lint]
extend-select = ["ANN"]
select = ["I"]
